6.16.20

+++ app/controllers/admin/boards_controller.rb
@@ -7,9 +7,11 @@ class Admin::BoardsController < ApplicationController
 
     def destroy
         board = Board.find(params[:id])
-        boardreport = Boardreport.find_by(board_id:params[:id])
+        if Boardreport.find_by(board_id:params[:id]).present?
+            boardreport = Boardreport.find_by(board_id:params[:id])
+            boardreport.destroy
+        end
         board.destroy
-        boardreport.destroy
         redirect_to request.referer
     end
